requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pypdf
python-docx
requests
python-decouple
google-genai
beautifulsoup4
lxml
aiofiles
httpx 